+++ changelog.md
@@ -1,3 +1,4 @@
 # 1.0.0
 - Core: copy all basics from ([Ewibot](https://github.com/Eccleria/ewibot)) project
 - Feat: add `stickers` handling for `MessageDelete` logs ([#15](https://github.com/Eccleria/Sil-Afian/issues/15)) ([#18](https://github.com/Eccleria/Sil-Afian/pull/18))
+- Update: change `activities` text so they are bot specific ([#23](https://github.com/Eccleria/Sil-Afian/pull/23))